backend/requirements.txt
sqlalchemy
fastapi
uvicorn
python-multipart
passlib
pyjwt
bcrypt==4.0.1